+++ packages/server/src/database/entities/ChatFlow.ts
@@ -1,9 +1,9 @@
 /* eslint-disable */
 import { Entity, Column, CreateDateColumn, UpdateDateColumn, PrimaryGeneratedColumn } from 'typeorm'
-import { ITool } from '../Interface'
+import { IChatFlow } from '../../Interface'
 
 @Entity()
-export class Tool implements ITool {
+export class ChatFlow implements IChatFlow {
     @PrimaryGeneratedColumn('uuid')
     id: string
 
@@ -11,19 +11,19 @@ export class Tool implements ITool {
     name: string
 
     @Column({ type: 'text' })
-    description: string
-
-    @Column()
-    color: string
+    flowData: string
 
     @Column({ nullable: true })
-    iconSrc?: string
+    deployed?: boolean
 
     @Column({ nullable: true })
-    schema?: string
+    isPublic?: boolean
 
     @Column({ nullable: true })
-    func?: string
+    apikeyid?: string
+
+    @Column({ nullable: true })
+    chatbotConfig?: string
 
     @CreateDateColumn()
     createdDate: Date
