+++ packages/server/src/database/entities/Credential.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
 import { Entity, Column, PrimaryGeneratedColumn, Index, CreateDateColumn, UpdateDateColumn } from 'typeorm'
-import { ICredential } from '../Interface'
+import { ICredential } from '../../Interface'
 
 @Entity()
 export class Credential implements ICredential {
