+++ packages/server/src/database/entities/index.ts
@@ -1,0 +1,11 @@
+import { ChatFlow } from './ChatFlow'
+import { ChatMessage } from './ChatMessage'
+import { Credential } from './Credential'
+import { Tool } from './Tool'
+
+export const entities = {
+    ChatFlow,
+    ChatMessage,
+    Credential,
+    Tool
+}
